+++ packages/csv-sort/rollup.config.js
@@ -1,59 +1,73 @@
-import resolve from 'rollup-plugin-node-resolve'
-import commonjs from 'rollup-plugin-commonjs'
-import uglify from 'rollup-plugin-uglify'
-import babel from 'rollup-plugin-babel'
-import { minify } from 'uglify-es'
-import pkg from './package.json'
+import resolve from "rollup-plugin-node-resolve";
+import commonjs from "rollup-plugin-commonjs";
+import uglify from "rollup-plugin-uglify";
+import strip from "rollup-plugin-strip";
+import babel from "rollup-plugin-babel";
+import { minify } from "uglify-es";
+import pkg from "./package.json";
 
-export default [
-  // browser-friendly UMD build
-  {
-    input: 'src/main.js',
-    output: {
-      file: pkg.browser,
-      format: 'umd',
-      name: 'csvSort',
+export default commandLineArgs => {
+  const finalConfig = [
+    // browser-friendly UMD build
+    {
+      input: "src/main.js",
+      output: {
+        file: pkg.browser,
+        format: "umd",
+        name: "csvSort"
+      },
+      plugins: [
+        strip({
+          sourceMap: false
+        }),
+        resolve(), // so Rollup can find deps
+        commonjs(), // so Rollup can convert deps to ES modules
+        babel(),
+        uglify({}, minify)
+      ]
     },
-    plugins: [
-      resolve(), // so Rollup can find deps
-      commonjs(), // so Rollup can convert deps to ES modules
-      babel(),
-      uglify({}, minify),
-    ],
-  },
 
-  // Builds: CommonJS (for Node) and ES module (for bundlers)
-  {
-    input: 'src/main.js',
-    output: [
-      { file: pkg.main, format: 'cjs' },
-      { file: pkg.module, format: 'es' },
-    ],
-    external: [
-      'bignumber.js',
-      'csv-split-easy',
-      'currency-symbol-map',
-      'is-numeric',
-      'lodash.pull',
-      'ordinal',
-    ],
-    plugins: [
-      babel(),
-    ],
-  },
+    // Builds: CommonJS (for Node) and ES module (for bundlers)
+    {
+      input: "src/main.js",
+      output: [
+        { file: pkg.main, format: "cjs" },
+        { file: pkg.module, format: "es" }
+      ],
+      external: [
+        "bignumber.js",
+        "csv-split-easy",
+        "is-numeric",
+        "lodash.pull",
+        "ordinal"
+      ],
+      plugins: [
+        strip({
+          sourceMap: false
+        }),
+        babel()
+      ]
+    },
 
-  // CommonJS util.js build in order to test
-  {
-    input: 'src/util.js',
-    output: [
-      { file: 'dist/util.cjs.js', format: 'cjs' },
-    ],
-    external: [
-      'currency-symbol-map/map',
-      'is-numeric',
-    ],
-    plugins: [
-      babel(),
-    ],
-  },
-]
+    // util.js needs transpiling as well:
+    {
+      input: "src/util.js",
+      output: [{ file: "dist/util.cjs.js", format: "cjs" }],
+      external: ["is-numeric"],
+      plugins: [
+        strip({
+          sourceMap: false
+        }),
+        babel()
+      ]
+    }
+  ];
+  if (commandLineArgs.dev) {
+    // if rollup was called without a --dev flag,
+    // dispose of a comment removal, strip():
+    finalConfig.forEach((singleConfigVal, i) => {
+      finalConfig[i].plugins.shift();
+    });
+  }
+  return finalConfig;
+};