+++ packages/csv-sort/src/main.js
@@ -1,40 +1,56 @@
 /* eslint prefer-destructuring:0, max-len:0 */
 
-import split from 'csv-split-easy'
-import pull from 'lodash.pull'
-import ordinal from 'ordinal'
-import BigNumber from 'bignumber.js'
-import findtype from './util'
+import split from "csv-split-easy";
+import pull from "lodash.pull";
+import ordinal from "ordinal";
+import BigNumber from "bignumber.js";
+import findtype from "./util";
 
-const isArr = Array.isArray
+const isArr = Array.isArray;
 
 function csvSort(input) {
-  let content
-  let msgContent = null
-  let msgType = null
+  let content;
+  let msgContent = null;
+  let msgType = null;
 
   // step 1.
   // ===========================
   // depends what was passed in,
-  if (typeof input === 'string') {
+  if (typeof input === "string") {
     if (input.length === 0) {
-      return [['']]
+      return [[""]];
     }
-    content = split(input)
+    content = split(input);
   } else if (isArr(input)) {
-    let culpritVal
-    let culpritIndex
-    if (!input.every((val, index) => {
-      if (!isArr(val)) {
-        culpritVal = val
-        culpritIndex = index
-      }
-      return isArr(val)
-    })) {
-      throw new TypeError(`csv-sort/csvSort(): [THROW_ID_01] the input is array as expected, but not all of its children are arrays! For example, the ${ordinal(culpritIndex)} element is not array but: ${typeof culpritVal}, equal to:\n${JSON.stringify(culpritVal, null, 4)}`)
+    let culpritVal;
+    let culpritIndex;
+    if (
+      !input.every((val, index) => {
+        if (!isArr(val)) {
+          culpritVal = val;
+          culpritIndex = index;
+        }
+        return isArr(val);
+      })
+    ) {
+      throw new TypeError(
+        `csv-sort/csvSort(): [THROW_ID_01] the input is array as expected, but not all of its children are arrays! For example, the ${ordinal(
+          culpritIndex
+        )} element is not array but: ${typeof culpritVal}, equal to:\n${JSON.stringify(
+          culpritVal,
+          null,
+          4
+        )}`
+      );
     }
   } else {
-    throw new TypeError(`csv-sort/csvSort(): [THROW_ID_02] The input is of a wrong type! We accept either string of array of arrays. We got instead: ${typeof input}, equal to:\n${JSON.stringify(input, null, 4)}`)
+    throw new TypeError(
+      `csv-sort/csvSort(): [THROW_ID_02] The input is of a wrong type! We accept either string of array of arrays. We got instead: ${typeof input}, equal to:\n${JSON.stringify(
+        input,
+        null,
+        4
+      )}`
+    );
   }
 
   // step 2.
@@ -44,26 +60,32 @@ function csvSort(input) {
   // - first row can have different amount of columns
   // - think about 2D trim feature
 
-  let schema = null
-  let stateHeaderRowPresent = false
-  let stateDataColumnRowLengthIsConsistent = true
-  const stateColumnsContainingSameValueEverywhere = []
+  let schema = null;
+  let stateHeaderRowPresent = false;
+  let stateDataColumnRowLengthIsConsistent = true;
+  const stateColumnsContainingSameValueEverywhere = [];
 
   // used for 2D trimming:
-  let indexAtWhichEmptyCellsStart = null
+  let indexAtWhichEmptyCellsStart = null;
 
   for (let i = content.length - 1; i >= 0; i--) {
     if (!schema) {
       // prevention against last blank row:
-      if ((content[i].length !== 1) || (content[i][0] !== '')) {
-        schema = []
+      if (content[i].length !== 1 || content[i][0] !== "") {
+        schema = [];
         for (let y = 0, len = content[i].length; y < len; y++) {
-          schema.push(findtype(content[i][y].trim()))
-          if ((indexAtWhichEmptyCellsStart === null) && (findtype(content[i][y].trim()) === 'empty')) {
-            indexAtWhichEmptyCellsStart = y
+          schema.push(findtype(content[i][y].trim()));
+          if (
+            indexAtWhichEmptyCellsStart === null &&
+            findtype(content[i][y].trim()) === "empty"
+          ) {
+            indexAtWhichEmptyCellsStart = y;
           }
-          if ((indexAtWhichEmptyCellsStart !== null) && (findtype(content[i][y].trim()) !== 'empty')) {
-            indexAtWhichEmptyCellsStart = null
+          if (
+            indexAtWhichEmptyCellsStart !== null &&
+            findtype(content[i][y].trim()) !== "empty"
+          ) {
+            indexAtWhichEmptyCellsStart = null;
           }
         }
       }
@@ -72,7 +94,9 @@ function csvSort(input) {
         // Check is this header row.
         // Header rows should consist of only text content.
         // Let's iterate through all elements and find out.
-        stateHeaderRowPresent = content[i].every(el => (findtype(el) === 'text') || (findtype(el) === 'empty'))
+        stateHeaderRowPresent = content[i].every(
+          el => findtype(el) === "text" || findtype(el) === "empty"
+        );
 
         // if schema was calculated (this means there's header row and at least one content row),
         // find out if the column length in the header differs from schema's
@@ -80,31 +104,40 @@ function csvSort(input) {
         // if (stateHeaderRowPresent && (schema.length !== content[i].length)) {
         // }
       }
-      if (!stateHeaderRowPresent && (schema.length !== content[i].length)) {
-        stateDataColumnRowLengthIsConsistent = false
+      if (!stateHeaderRowPresent && schema.length !== content[i].length) {
+        stateDataColumnRowLengthIsConsistent = false;
       }
 
-      let perRowIndexAtWhichEmptyCellsStart = null
+      let perRowIndexAtWhichEmptyCellsStart = null;
       for (let y = 0, len = content[i].length; y < len; y++) {
         // trim
-        if ((perRowIndexAtWhichEmptyCellsStart === null) && (findtype(content[i][y].trim()) === 'empty')) {
-          perRowIndexAtWhichEmptyCellsStart = y
+        if (
+          perRowIndexAtWhichEmptyCellsStart === null &&
+          findtype(content[i][y].trim()) === "empty"
+        ) {
+          perRowIndexAtWhichEmptyCellsStart = y;
         }
-        if ((perRowIndexAtWhichEmptyCellsStart !== null) && (findtype(content[i][y].trim()) !== 'empty')) {
-          perRowIndexAtWhichEmptyCellsStart = null
+        if (
+          perRowIndexAtWhichEmptyCellsStart !== null &&
+          findtype(content[i][y].trim()) !== "empty"
+        ) {
+          perRowIndexAtWhichEmptyCellsStart = null;
         }
         // checking schema
-        if ((findtype(content[i][y].trim()) !== schema[y]) && !stateHeaderRowPresent) {
-          const toAdd = findtype(content[i][y].trim())
+        if (
+          findtype(content[i][y].trim()) !== schema[y] &&
+          !stateHeaderRowPresent
+        ) {
+          const toAdd = findtype(content[i][y].trim());
           if (isArr(schema[y])) {
             if (!schema[y].includes(toAdd)) {
-              schema[y].push(findtype(content[i][y].trim()))
+              schema[y].push(findtype(content[i][y].trim()));
             }
           } else if (schema[y] !== toAdd) {
-            const temp = schema[y]
-            schema[y] = []
-            schema[y].push(temp)
-            schema[y].push(toAdd)
+            const temp = schema[y];
+            schema[y] = [];
+            schema[y].push(temp);
+            schema[y].push(toAdd);
           }
         } // else {
         // }
@@ -126,34 +159,36 @@ function csvSort(input) {
       //
 
       if (
-        (indexAtWhichEmptyCellsStart !== null) &&
-        (perRowIndexAtWhichEmptyCellsStart !== null) &&
-        (perRowIndexAtWhichEmptyCellsStart > indexAtWhichEmptyCellsStart) &&
-        (!stateHeaderRowPresent || (stateHeaderRowPresent && (i !== 0)))
+        indexAtWhichEmptyCellsStart !== null &&
+        perRowIndexAtWhichEmptyCellsStart !== null &&
+        perRowIndexAtWhichEmptyCellsStart > indexAtWhichEmptyCellsStart &&
+        (!stateHeaderRowPresent || (stateHeaderRowPresent && i !== 0))
       ) {
-        indexAtWhichEmptyCellsStart = perRowIndexAtWhichEmptyCellsStart
+        indexAtWhichEmptyCellsStart = perRowIndexAtWhichEmptyCellsStart;
       }
     }
   }
 
   if (!indexAtWhichEmptyCellsStart) {
-    indexAtWhichEmptyCellsStart = schema.length
+    indexAtWhichEmptyCellsStart = schema.length;
   }
 
   // find out at which index non-empty columns start. This is effectively left-side trimming.
-  let nonEmptyColsStartAt = 0
+  let nonEmptyColsStartAt = 0;
   for (let i = 0, len = schema.length; i < len; i++) {
-    if (schema[i] === 'empty') {
-      nonEmptyColsStartAt = i
+    if (schema[i] === "empty") {
+      nonEmptyColsStartAt = i;
     } else {
-      break
+      break;
     }
   }
 
   // if there are empty column in front, trim (via slice) both content and schema
   if (nonEmptyColsStartAt !== 0) {
-    content = content.map(arr => arr.slice(nonEmptyColsStartAt + 1, indexAtWhichEmptyCellsStart))
-    schema = schema.slice(nonEmptyColsStartAt + 1, indexAtWhichEmptyCellsStart)
+    content = content.map(arr =>
+      arr.slice(nonEmptyColsStartAt + 1, indexAtWhichEmptyCellsStart)
+    );
+    schema = schema.slice(nonEmptyColsStartAt + 1, indexAtWhichEmptyCellsStart);
   }
 
   // step 3.
@@ -169,21 +204,23 @@ function csvSort(input) {
 
   // swoop in traversing the schema array to get "numeric" columns:
   // ----------------
-  const numericSchemaColumns = []
-  let balanceColumnIndex
+  const numericSchemaColumns = [];
+  let balanceColumnIndex;
   schema.forEach((colType, i) => {
-    if (colType === 'numeric') {
-      numericSchemaColumns.push(i)
+    if (colType === "numeric") {
+      numericSchemaColumns.push(i);
     }
-  })
+  });
 
-  const traverseUpToThisIndexAtTheTop = stateHeaderRowPresent ? 1 : 0
+  const traverseUpToThisIndexAtTheTop = stateHeaderRowPresent ? 1 : 0;
 
   if (numericSchemaColumns.length === 1) {
     // Bob's your uncle, the only numeric column is your Balance column
-    balanceColumnIndex = numericSchemaColumns[0]
+    balanceColumnIndex = numericSchemaColumns[0];
   } else if (numericSchemaColumns.length === 0) {
-    throw new Error('csv-sort/csvSort(): [THROW_ID_03] Your CSV file does not contain numeric-only columns and computer was not able to detect the "Balance" column!')
+    throw new Error(
+      'csv-sort/csvSort(): [THROW_ID_03] Your CSV file does not contain numeric-only columns and computer was not able to detect the "Balance" column!'
+    );
   } else {
     // So (numericSchemaColumns > 0) and we'll have to do some work.
     // Fine.
@@ -197,13 +234,18 @@ function csvSort(input) {
     // find out `balanceColumnIndex`:
     // ----------------
 
-    let potentialBalanceColumnIndexesList = Array.from(numericSchemaColumns)
+    let potentialBalanceColumnIndexesList = Array.from(numericSchemaColumns);
     // iterate through `potentialBalanceColumnIndexesList`
-    const deleteFromPotentialBalanceColumnIndexesList = []
+    const deleteFromPotentialBalanceColumnIndexesList = [];
 
-    for (let i = 0, len = potentialBalanceColumnIndexesList.length; i < len; i++) {
+    for (
+      let i = 0, len = potentialBalanceColumnIndexesList.length;
+      i < len;
+      i++
+    ) {
       // if any two rows are in sequence currently and they are equal, this column is out
-      const suspectedBalanceColumnsIndexNumber = potentialBalanceColumnIndexesList[i]
+      const suspectedBalanceColumnsIndexNumber =
+        potentialBalanceColumnIndexesList[i];
       // we traverse column suspected to be "Balance" with index `index` vertically,
       // from the top to bottom. Depending if there's heading row, we start at 0 or 1,
       // which is set by `traverseUpToThisIndexAtTheTop`.
@@ -212,10 +254,10 @@ function csvSort(input) {
 
       // EASY ATTEMPT TO RULE-OUT NOT-BALANCE COLUMNS
 
-      let previousValue // to check if two consecutive are the same
-      let lookForTwoEqualAndConsecutive = true
-      let firstValue // to check if all are the same
-      let lookForAllTheSame = true
+      let previousValue; // to check if two consecutive are the same
+      let lookForTwoEqualAndConsecutive = true;
+      let firstValue; // to check if all are the same
+      let lookForAllTheSame = true;
 
       for (
         let rowNum = traverseUpToThisIndexAtTheTop, len2 = content.length;
@@ -225,50 +267,62 @@ function csvSort(input) {
         // 1. check for two consecutive equal values
         if (lookForTwoEqualAndConsecutive) {
           if (previousValue === undefined) {
-            previousValue = content[rowNum][suspectedBalanceColumnsIndexNumber]
-          } else if (previousValue === content[rowNum][suspectedBalanceColumnsIndexNumber]) {
+            previousValue = content[rowNum][suspectedBalanceColumnsIndexNumber];
+          } else if (
+            previousValue ===
+            content[rowNum][suspectedBalanceColumnsIndexNumber]
+          ) {
             // potentialBalanceColumnIndexesList.splice(suspectedBalanceColumnsIndexNumber, 1)
             // don't mutate the `potentialBalanceColumnIndexesList`, do it later.
             // Let's compile TO-DELETE list instead:
-            deleteFromPotentialBalanceColumnIndexesList.push(suspectedBalanceColumnsIndexNumber)
-            lookForTwoEqualAndConsecutive = false
+            deleteFromPotentialBalanceColumnIndexesList.push(
+              suspectedBalanceColumnsIndexNumber
+            );
+            lookForTwoEqualAndConsecutive = false;
           } else {
-            previousValue = content[rowNum][suspectedBalanceColumnsIndexNumber]
+            previousValue = content[rowNum][suspectedBalanceColumnsIndexNumber];
           }
         }
         // 2. also, tell if ALL values are the same:
         if (lookForAllTheSame) {
           if (firstValue === undefined) {
-            firstValue = content[rowNum][suspectedBalanceColumnsIndexNumber]
-          } else if (content[rowNum][suspectedBalanceColumnsIndexNumber] !== firstValue) {
-            lookForAllTheSame = false
+            firstValue = content[rowNum][suspectedBalanceColumnsIndexNumber];
+          } else if (
+            content[rowNum][suspectedBalanceColumnsIndexNumber] !== firstValue
+          ) {
+            lookForAllTheSame = false;
           }
         }
         if (!lookForTwoEqualAndConsecutive) {
-          break
+          break;
         }
       }
 
       if (lookForAllTheSame) {
-        stateColumnsContainingSameValueEverywhere.push(suspectedBalanceColumnsIndexNumber)
+        stateColumnsContainingSameValueEverywhere.push(
+          suspectedBalanceColumnsIndexNumber
+        );
       }
     }
 
     // now mutate the `potentialBalanceColumnIndexesList` using
     // `deleteFromPotentialBalanceColumnIndexesList`:
-    potentialBalanceColumnIndexesList = pull(potentialBalanceColumnIndexesList, ...deleteFromPotentialBalanceColumnIndexesList)
+    potentialBalanceColumnIndexesList = pull(
+      potentialBalanceColumnIndexesList,
+      ...deleteFromPotentialBalanceColumnIndexesList
+    );
 
     if (potentialBalanceColumnIndexesList.length === 1) {
-      balanceColumnIndex = potentialBalanceColumnIndexesList[0]
+      balanceColumnIndex = potentialBalanceColumnIndexesList[0];
     } else if (potentialBalanceColumnIndexesList.length === 0) {
-      throw new Error('csv-sort/csvSort(): [THROW_ID_04] The computer can\'t find the "Balance" column! It saw some numeric-only columns, but they all seem to have certain rows with the same values as rows right below/above them!')
+      throw new Error(
+        'csv-sort/csvSort(): [THROW_ID_04] The computer can\'t find the "Balance" column! It saw some numeric-only columns, but they all seem to have certain rows with the same values as rows right below/above them!'
+      );
     } else {
       // TODO - continue processing interpolating horizontally and vertically.
       //
       //
-
       // COMPLEX ATTEMPT TO RULE-OUT NOT-BALANCE COLUMNS
-
       //
       //
       // zzz
@@ -282,7 +336,9 @@ function csvSort(input) {
   }
 
   if (!balanceColumnIndex) {
-    throw new Error('csv-sort/csvSort(): [THROW_ID_05] Sadly computer couldn\'t find its way in this CSV and had to stop working on it.')
+    throw new Error(
+      "csv-sort/csvSort(): [THROW_ID_05] Sadly computer couldn't find its way in this CSV and had to stop working on it."
+    );
   }
 
   // step 4.
@@ -291,17 +347,26 @@ function csvSort(input) {
 
   // take schema, filter all indexes that are equal to or are arrays and have
   // "numeric" among their values, then remove the index of "Balance" column:
-  const potentialCreditDebitColumns = pull(Array.from(schema.reduce((result, el, index) => {
-    if (((typeof el === 'string') && el === 'numeric') || (isArr(el) && el.includes('numeric'))) {
-      result.push(index)
-    }
-    return result
-  }, [])), balanceColumnIndex, ...stateColumnsContainingSameValueEverywhere)
+  const potentialCreditDebitColumns = pull(
+    Array.from(
+      schema.reduce((result, el, index) => {
+        if (
+          (typeof el === "string" && el === "numeric") ||
+          (isArr(el) && el.includes("numeric"))
+        ) {
+          result.push(index);
+        }
+        return result;
+      }, [])
+    ),
+    balanceColumnIndex,
+    ...stateColumnsContainingSameValueEverywhere
+  );
 
   // step 5.
   // ===========================
 
-  const resContent = []
+  const resContent = [];
 
   // Now that we know the `balanceColumnIndex`, traverse the CSV rows again,
   // assembling a new array
@@ -311,78 +376,127 @@ function csvSort(input) {
   // Worst case scenario, if it doesn't match with anything, we'll throw in the end.
   // For now, let's assume CSV is correct, only rows are mixed.
 
-  resContent.push(content[content.length - 1].slice(0, indexAtWhichEmptyCellsStart))
+  resContent.push(
+    content[content.length - 1].slice(0, indexAtWhichEmptyCellsStart)
+  );
 
-  const usedUpRows = []
+  const usedUpRows = [];
 
-  const bottom = stateHeaderRowPresent ? 1 : 0
+  const bottom = stateHeaderRowPresent ? 1 : 0;
   for (let y = content.length - 2; y >= bottom; y--) {
     // for each row above the last-one (which is already in place), we'll traverse
     // all the rows above to find the match.
     // go through all the rows and pick the right row which matches to the above:
-    for (let suspectedRowsIndex = content.length - 2; suspectedRowsIndex >= bottom; suspectedRowsIndex--) {
+    for (
+      let suspectedRowsIndex = content.length - 2;
+      suspectedRowsIndex >= bottom;
+      suspectedRowsIndex--
+    ) {
       if (!usedUpRows.includes(suspectedRowsIndex)) {
         // go through each of the suspected Credit/Debit columns:
 
-        let thisRowIsDone = false
+        let thisRowIsDone = false;
 
-        for (let suspectedColIndex = 0, len = potentialCreditDebitColumns.length; suspectedColIndex < len; suspectedColIndex++) {
-          let diffVal = null
-          if (content[suspectedRowsIndex][potentialCreditDebitColumns[suspectedColIndex]] !== '') {
-            diffVal = new BigNumber(content[suspectedRowsIndex][potentialCreditDebitColumns[suspectedColIndex]])
+        for (
+          let suspectedColIndex = 0, len = potentialCreditDebitColumns.length;
+          suspectedColIndex < len;
+          suspectedColIndex++
+        ) {
+          let diffVal = null;
+          if (
+            content[suspectedRowsIndex][
+              potentialCreditDebitColumns[suspectedColIndex]
+            ] !== ""
+          ) {
+            diffVal = new BigNumber(
+              content[suspectedRowsIndex][
+                potentialCreditDebitColumns[suspectedColIndex]
+              ]
+            );
           }
 
-          let totalVal = null
-          if (content[suspectedRowsIndex][balanceColumnIndex] !== '') {
-            totalVal = new BigNumber(content[suspectedRowsIndex][balanceColumnIndex])
+          let totalVal = null;
+          if (content[suspectedRowsIndex][balanceColumnIndex] !== "") {
+            totalVal = new BigNumber(
+              content[suspectedRowsIndex][balanceColumnIndex]
+            );
           }
 
-          let topmostResContentBalance = null
-          if (resContent[0][balanceColumnIndex] !== '') {
-            topmostResContentBalance = new BigNumber(resContent[0][balanceColumnIndex])
+          let topmostResContentBalance = null;
+          if (resContent[0][balanceColumnIndex] !== "") {
+            topmostResContentBalance = new BigNumber(
+              resContent[0][balanceColumnIndex]
+            );
           }
 
-          let currentRowsDiffVal = null
-          if (resContent[resContent.length - 1][potentialCreditDebitColumns[suspectedColIndex]] !== '') {
-            currentRowsDiffVal = new BigNumber(resContent[resContent.length - 1][potentialCreditDebitColumns[suspectedColIndex]])
+          let currentRowsDiffVal = null;
+          if (
+            resContent[resContent.length - 1][
+              potentialCreditDebitColumns[suspectedColIndex]
+            ] !== ""
+          ) {
+            currentRowsDiffVal = new BigNumber(
+              resContent[resContent.length - 1][
+                potentialCreditDebitColumns[suspectedColIndex]
+              ]
+            );
           }
 
-          let lastResContentRowsBalance = null
-          if (resContent[resContent.length - 1][balanceColumnIndex] !== '') {
-            lastResContentRowsBalance = new BigNumber(resContent[resContent.length - 1][balanceColumnIndex])
+          let lastResContentRowsBalance = null;
+          if (resContent[resContent.length - 1][balanceColumnIndex] !== "") {
+            lastResContentRowsBalance = new BigNumber(
+              resContent[resContent.length - 1][balanceColumnIndex]
+            );
           }
 
-          if (diffVal && (totalVal.plus(diffVal).eq(topmostResContentBalance))) {
+          if (diffVal && totalVal.plus(diffVal).eq(topmostResContentBalance)) {
             // ADD THIS ROW ABOVE EVERYTHING
             // add this row above the current HEAD in resContent lines array (index `0`)
-            resContent.unshift(content[suspectedRowsIndex].slice(0, indexAtWhichEmptyCellsStart))
-            usedUpRows.push(suspectedRowsIndex)
-            thisRowIsDone = true
-            break
-          } else if (diffVal && (totalVal.minus(diffVal).eq(topmostResContentBalance))) {
+            resContent.unshift(
+              content[suspectedRowsIndex].slice(0, indexAtWhichEmptyCellsStart)
+            );
+            usedUpRows.push(suspectedRowsIndex);
+            thisRowIsDone = true;
+            break;
+          } else if (
+            diffVal &&
+            totalVal.minus(diffVal).eq(topmostResContentBalance)
+          ) {
             // ADD THIS ROW ABOVE EVERYTHING
-            resContent.unshift(content[suspectedRowsIndex].slice(0, indexAtWhichEmptyCellsStart))
-            usedUpRows.push(suspectedRowsIndex)
-            thisRowIsDone = true
-            break
-          } else if (currentRowsDiffVal && (lastResContentRowsBalance.plus(currentRowsDiffVal).eq(totalVal))) {
+            resContent.unshift(
+              content[suspectedRowsIndex].slice(0, indexAtWhichEmptyCellsStart)
+            );
+            usedUpRows.push(suspectedRowsIndex);
+            thisRowIsDone = true;
+            break;
+          } else if (
+            currentRowsDiffVal &&
+            lastResContentRowsBalance.plus(currentRowsDiffVal).eq(totalVal)
+          ) {
             // ADD THIS ROW BELOW EVERYTHING
-            resContent.push(content[suspectedRowsIndex].slice(0, indexAtWhichEmptyCellsStart))
-            usedUpRows.push(suspectedRowsIndex)
-            thisRowIsDone = true
-            break
-          } else if (currentRowsDiffVal && (lastResContentRowsBalance.minus(currentRowsDiffVal).eq(totalVal))) {
+            resContent.push(
+              content[suspectedRowsIndex].slice(0, indexAtWhichEmptyCellsStart)
+            );
+            usedUpRows.push(suspectedRowsIndex);
+            thisRowIsDone = true;
+            break;
+          } else if (
+            currentRowsDiffVal &&
+            lastResContentRowsBalance.minus(currentRowsDiffVal).eq(totalVal)
+          ) {
             // ADD THIS ROW BELOW EVERYTHING
-            resContent.push(content[suspectedRowsIndex].slice(0, indexAtWhichEmptyCellsStart))
-            usedUpRows.push(suspectedRowsIndex)
-            thisRowIsDone = true
-            break
+            resContent.push(
+              content[suspectedRowsIndex].slice(0, indexAtWhichEmptyCellsStart)
+            );
+            usedUpRows.push(suspectedRowsIndex);
+            thisRowIsDone = true;
+            break;
           } // else {
           // }
         }
         if (thisRowIsDone) {
-          thisRowIsDone = false
-          break
+          thisRowIsDone = false;
+          break;
         }
       }
     }
@@ -391,23 +505,26 @@ function csvSort(input) {
   // restore title row if present
   if (stateHeaderRowPresent) {
     // trim header row of trailing empty columns if they protrude outside of the (consistent row length) schema
-    if (stateDataColumnRowLengthIsConsistent && (content[0].length > schema.length)) {
-      content[0].length = schema.length
+    if (
+      stateDataColumnRowLengthIsConsistent &&
+      content[0].length > schema.length
+    ) {
+      content[0].length = schema.length;
     }
     // push header row on top of the results array:
-    resContent.unshift(content[0].slice(0, indexAtWhichEmptyCellsStart))
+    resContent.unshift(content[0].slice(0, indexAtWhichEmptyCellsStart));
   }
 
-  if ((content.length - (stateHeaderRowPresent ? 2 : 1)) !== usedUpRows.length) {
-    msgContent = 'Not all rows were recognised!'
-    msgType = 'alert'
+  if (content.length - (stateHeaderRowPresent ? 2 : 1) !== usedUpRows.length) {
+    msgContent = "Not all rows were recognised!";
+    msgType = "alert";
   }
 
   return {
     res: resContent,
     msgContent,
-    msgType,
-  }
+    msgType
+  };
 }
 
-export { csvSort as default }
+export { csvSort as default };